+++ R/yeojohnson.R
@@ -114,8 +114,8 @@ print.yeojohnson <- function(x, ...) {
       'Yeo-Johnson Transformation with', x$n, 'nonmissing obs.:\n', 
       'Estimated statistics:\n',
       '- lambda =', x$lambda, '\n',
-      '- mean =', x$mean, '\n',
-      '- sd =', x$sd, '\n')
+      '- mean (post-transform) =', x$mean, '\n',
+      '- sd (post-transform) =', x$sd, '\n')
 }
 
 # Helper functions that estimates yj lambda parameter